# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /payments/checkout:
    post:
      operationId: createPaymentCheckout
      tags: [payments]
      summary: Create an idempotent payment checkout
      description: Creates a pending checkout. A payment only affects the ledger after a verified provider event is received. Test-mode checkouts are completed by a server-generated test event.
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentCheckoutInput"
      responses:
        "201":
          description: Payment checkout
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
        "400":
          description: Invalid checkout or idempotency key
        "401":
          description: Authentication required
        "409":
          description: Idempotency key conflict
        "503":
          description: Live payments are not enabled or configured
  /payments/{id}:
    get:
      operationId: getPayment
      tags: [payments]
      summary: Get one of the authenticated member's payments
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Payment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
        "401":
          description: Authentication required
        "403":
          description: Payment is not owned by the member
        "404":
          description: Payment not found
  /payments/{id}/refund:
    post:
      operationId: refundPayment
      tags: [payments]
      summary: Request a refund for an eligible payment
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentRefundInput"
      responses:
        "200":
          description: Refund and updated payment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentRefundResponse"
        "400":
          description: Invalid refund
        "401":
          description: Authentication required
        "403":
          description: Payment is not owned by the member
        "404":
          description: Payment not found
        "409":
          description: Payment is not refundable or amount exceeds the balance
        "503":
          description: Live refunds are not enabled or configured
  /paypal/create-order:
    post:
      operationId: createPayPalOrder
      tags: [payments]
      summary: Create a PayPal Orders v2 order for a local payment
      description: Authenticated. Accepts only the opaque local payment ID; amount and currency are derived server-side.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PayPalCreateOrderInput"
      responses:
        "201": { description: PayPal order created }
  /paypal/capture-order:
    post:
      operationId: capturePayPalOrder
      tags: [payments]
      summary: Capture an approved PayPal order
      description: Authenticated. The server verifies capture status, local order binding, USD currency, and cents before finality.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PayPalCaptureOrderInput"
      responses:
        "200": { description: Verified payment result }
  /paypal/webhook:
    post:
      operationId: receivePayPalWebhook
      tags: [payments]
      summary: Receive a verified PayPal webhook
      description: Raw JSON only. The server verifies PayPal transmission headers through PayPal's verify-webhook-signature endpoint.
      responses:
        "202": { description: Event accepted }
        "401": { description: Invalid webhook signature }
  /payments/webhooks/{provider}:
    post:
      operationId: receivePaymentWebhook
      tags: [payments]
      summary: Receive a signed provider payment event
      description: Provider events must include a valid HMAC signature. Replayed event IDs are acknowledged without applying a second ledger mutation.
      parameters:
        - name: provider
          in: path
          required: true
          schema: { type: string, minLength: 1 }
        - name: x-payment-signature
          in: header
          required: true
          schema: { type: string, minLength: 1 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProviderWebhookPayload"
      responses:
        "200":
          description: Event accepted and reconciled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentWebhookResponse"
        "400":
          description: Invalid or unsupported event
        "401":
          description: Invalid webhook signature
        "404":
          description: Payment not found
  /admin/payments/reconciliation:
    post:
      operationId: reconcilePayments
      tags: [moderation]
      summary: Reconcile stale payment states
      description: Marks pending checkouts older than fifteen minutes for manual provider reconciliation. It never fabricates a successful payment.
      responses:
        "200":
          description: Reconciliation report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentReconciliation"
        "401":
          description: Moderator identity required
        "403":
          description: Moderator role required
...
components:
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: A stable key for safely retrying the same payment operation.
      schema:
        type: string
        minLength: 1
        maxLength: 255
...
  schemas:
    PayPalCreateOrderInput:
      type: object
      required: [paymentId]
      properties:
        paymentId: { type: string }
    PayPalCaptureOrderInput:
      type: object
      required: [paymentId, orderId]
      properties:
        paymentId: { type: string }
        orderId: { type: string }
...
    PaymentCheckoutInput:
      type: object
      properties:
        listingId: { type: string, minLength: 1 }
        type:
          type: string
          enum: [OWNER_BID, COMMUNITY_BID, PENALTY]
        amount: { type: number, exclusiveMinimum: 0, maximum: 100000 }
        currency: { type: string, minLength: 3, maxLength: 3, default: USD }
        reason: { type: string, maxLength: 120 }
        receiptEmail: { type: ["string", "null"], maxLength: 320 }
      required: [listingId, type, amount]
    PaymentRefundInput:
      type: object
      properties:
        amount: { type: number, exclusiveMinimum: 0, maximum: 100000 }
        reason: { type: string, maxLength: 500 }
      required: [amount]
    Payment:
      type: object
      properties:
        id: { type: string }
        listingId:
          type: ["string", "null"]
          description: Null until a deferred new-listing checkout's payment succeeds and its listing is created.
        type: { type: string, enum: [OWNER_BID, COMMUNITY_BID, PENALTY, SPONSORSHIP] }
        amount: { type: number }
        currency: { type: string }
        platformFee: { type: number }
        provider: { type: string }
        status:
          type: string
          enum: [pending, succeeded, failed, refunded, partially_refunded, requires_reconciliation]
        providerCheckoutId: { type: ["string", "null"] }
        providerPaymentId: { type: ["string", "null"] }
        failureCode: { type: ["string", "null"] }
        failureMessage: { type: ["string", "null"] }
        receiptNumber: { type: ["string", "null"] }
        receiptEmail: { type: ["string", "null"] }
        checkoutUrl:
          type: ["string", "null"]
          description: Hosted provider sandbox checkout URL when an external checkout is pending.
        createdAt: { type: string }
        updatedAt: { type: string }
      required: [id, listingId, type, amount, currency, platformFee, provider, status, providerCheckoutId, providerPaymentId, failureCode, failureMessage, receiptNumber, receiptEmail, createdAt, updatedAt]
    PaymentRefund:
      type: object
      properties:
        id: { type: string }
        paymentId: { type: string }
        amount: { type: number }
        reason: { type: ["string", "null"] }
        status: { type: string, enum: [pending, succeeded, failed] }
        providerRefundId: { type: ["string", "null"] }
        createdAt: { type: string }
      required: [id, paymentId, amount, reason, status, providerRefundId, createdAt]
    PaymentRefundResponse:
      type: object
      properties:
        refund: { $ref: "#/components/schemas/PaymentRefund" }
        payment: { $ref: "#/components/schemas/Payment" }
      required: [refund, payment]
    PaymentWebhookResponse:
      type: object
      properties:
        eventId: { type: string }
        duplicate: { type: boolean }
        handled: { type: boolean }
        payment: { $ref: "#/components/schemas/Payment" }
      required: [eventId, duplicate, handled, payment]
    ProviderWebhookPayload:
      type: object
      description: Provider-specific signed event payload.
      additionalProperties: true
    PaymentReconciliation:
      type: object
      properties:
        markedForReview: { type: number }
        pending: { type: number }
        failed: { type: number }
        succeeded: { type: number }
      required: [markedForReview, pending, failed, succeeded]
...